+++ src/golden/config.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+"""Config: the golden girl
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+
+# configs
+CHARAS = (
+        )
+
+STAGES = (
+        )
+
+DAYS = (
+        )
+
+ITEMS = (
+        )
+
+INFOS = (
+        )
+
+FLAGS = (
+        )
+
+THEMES = {
+        }