+++ src/golden/story.py
@@ -1,0 +1,38 @@
+# -*- coding: utf-8 -*-
+"""Story: the golden girl
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+from storybuilder.builder import world as wd
+from src.golden import config as cnf
+
+
+# episodes
+def ep_intro(w: wd.World):
+    return []
+
+# main
+def world():
+    w = wd.World("The golden girld")
+    w.set_db(cnf.CHARAS, cnf.STAGES, cnf.DAYS, cnf.ITEMS, cnf.INFOS, cnf.FLAGS)
+    return w
+
+
+def story(w: wd.World):
+    return (w.maintitle("不幸の黄金少女"),
+            ep_intro(w),
+            )
+
+
+def main(): # pragma: no cover
+    w = world()
+    return w.build(story(w))
+
+
+if __name__ == '__main__':
+    import sys
+    sys.exit(main())
+